Refactor sparse package structure
