+++ xobjects/sparse/__init__.py
@@ -1,3 +1,3 @@
-from .sparse import factorized_sparse_solver
+from ._sparse import factorized_sparse_solver
 from . import solvers
 __all__ = ["factorized_sparse_solver","solvers"]
