+++ xobjects/sparse/solvers/CUDA/__init__.py
@@ -2,12 +2,15 @@ __all__ = []
 try:
     from ._cuDSSLU import DirectSolverSuperLU as cuDSSSuperLU
     __all__.append("cuDSSSuperLU")
-except:
+except (ModuleNotFoundError,ImportError):
     pass
 try:
     from ._luLU import luLU as CachedSuperLU
     __all__.append("CachedSuperLU")
     from cupyx.scipy.sparse.linalg import splu as CupySuperLU
     __all__.append("CupySuperLU")
-except:
+except (ModuleNotFoundError,ImportError):
     pass
+
+if not __all__:
+    raise ImportError("Failed to import any CUDA-based sparse solver")
