+++ xobjects/sparse/solvers/__init__.py
@@ -3,5 +3,5 @@ __all__ = ["CPU"]
 try:
     from . import CUDA
     __all__.append("CUDA")
-except:
+except ImportError:
     pass
